consumer/requirements.txt
aiohttp==3.7.4.post0
aiohttp-sse-client==0.2.1
aioredis==1.3.1
async-timeout==3.0.1; python_full_version >= "3.5.3"
attrs==20.3.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
blaseball-mike==5.9.2
certifi==2020.12.5
chardet==4.0.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
dateparser==1.0.0
hiredis==2.0.0; python_version >= "3.6"
idna==2.10; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
multidict==5.1.0; python_version >= "3.6"
python-dateutil==2.8.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
pytz==2021.1
regex==2021.4.4
requests==2.25.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
requests-cache==0.5.2
six==1.15.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2; python_version >= "3.6"
unpaddedbase64==2.1.0
urllib3==1.26.4; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4" and python_version < "4"
yarl==1.6.3; python_version >= "3.6"
